+++ projects/js-upload-api/src/index.ts
@@ -1,6 +1,6 @@
 "use strict";
 export { Client } from "./Client.js";
-export { ClientPKey } from "./ClientPkey.js";
+export { ClientPKey } from "./ClientPKey.js";
 export { ClientType } from "./types.js";
 export { FileType, File, EdgeFile, NodeFile } from "./File.js";
 export { Dataset } from "./Dataset.js";
